// nextjs/src/types/grant.ts
type Grant = {
// ...
  collabOpportunitySubscribers: Initiative[];

  /**
   * True iff this Grant has been recommended by NSR; false otherwise.
   */
  isRecommended: boolean;
};